certvalidator/ml/requirements.txt
# Core DL stack — CUDA 12.1 build (RTX 3050, driver >= 525)
# Install with: pip install torch==2.2.2+cu121 torchvision==0.17.2+cu121 torchaudio==2.2.2+cu121 --index-url https://download.pytorch.org/whl/cu121
torch==2.2.2+cu121
torchvision==0.17.2+cu121
torchaudio==2.2.2+cu121

# HuggingFace ecosystem
transformers==4.40.0
datasets==2.19.0
tokenizers==0.19.1
accelerate==0.29.3
evaluate==0.4.1

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# Image processing
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
Pillow==10.3.0
scikit-image==0.23.1
albumentations==1.4.4
imgaug==0.4.0

# PDF handling
PyMuPDF==1.24.1
pdf2image==1.17.0
pypdf==4.2.0

# Data science
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2

# Annotation / visualisation
supervision==0.19.0

# Synthetic data generation
faker==24.11.0
reportlab==4.1.0
fpdf2==2.7.9

# Model utilities
timm==1.0.3
einops==0.8.0
onnx==1.16.0
onnxruntime-gpu==1.20.0

# Experiment tracking
mlflow==2.12.1

# Utilities
tqdm==4.66.2
pyyaml==6.0.1
python-dotenv==1.0.1
rich==13.7.1
typer==0.12.3
